backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
celery[redis]
redis
gitpython
anthropic
httpx
weasyprint
apscheduler
sse-starlette
python-multipart
aiofiles
lizard
radon
bandit
ruff
vulture
pip-audit
detect-secrets
checkov
pytest
pytest-asyncio
